requirements.txt
flask
python-dotenv
flask-smorest
sqlalchemy
flask-sqlalchemy
flask-jwt-extended
passlib
flask-migrate
gunicorn
requests
rq
redis